refactor: move the model base classes and site models to the platform
Two related moves, the first a prerequisite for the second.

AppBase, HistoryMixin and HistoryAwareManager were generic infrastructure
living in n23/core/models/ — 67 references, and the base class of nearly
every edition model. They are now gyrinx/base_models.py,
gyrinx/history_mixin.py and gyrinx/history_aware_manager.py. All three are
abstract, so this is a pure Python move with no migration. It also unblocks
platform models inheriting AppBase, which is what the rest of this commit
needs.

Banner and ImpersonationLog then move to a new platform app,
gyrinx.site (label `gyrinxsite`) — an announcement banner and an
audit log of admin impersonation are about running Gyrinx, not about
playing a game edition. Same technique as the accounts move: db_table
pinned to core_banner / core_historicalbanner / core_impersonationlog,
migrations wrapped in SeparateDatabaseAndState with no database
operations, and the ContentType rows carried across in place so attached
permissions follow.

Notification does NOT move, despite being the third model the platform
wanted. It declares ForeignKeys to core.List and core.Campaign, so
relocating it would give a platform model hard references to edition
tables — replacing the inversion rather than removing it. Generalising
those two columns is a design question, not a mechanical one, so it stays
put along with the notifications context processor.

Verified against a populated database: the banner row survives, the
ContentType rows move label while keeping their ids, and no gyrinxsite_*
table is created. 4166 tests pass, makemigrations --check clean.

Refs #2093
