+++ gyrinx/history_aware_manager.py
