+++ gyrinx/settings.py
@@ -156,6 +156,7 @@ INSTALLED_APPS = [
     "n23.core",
     "n23.content",
     "gyrinx.accounts",
+    "gyrinx.site",
     "gyrinx.analytics",
     # Must be loaded AFTER gyrinx.analytics so MaintenanceAdminSite extends
     # AnalyticsAdminSite (see gyrinx/maintenance/admin.py).
