+++ gyrinx/site/apps.py
@@ -1,0 +1,7 @@
+from django.apps import AppConfig
+
+
+class SiteConfig(AppConfig):
+    default_auto_field = "django.db.models.BigAutoField"
+    name = "gyrinx.site"
+    label = "gyrinxsite"
