+++ gyrinx/site/models.py
@@ -2,7 +2,7 @@ from django.core.cache import cache
 from django.db import models
 from simple_history.models import HistoricalRecords
 
-from n23.core.models.base import AppBase
+from gyrinx.base_models import AppBase
 
 
 class Banner(AppBase):
@@ -41,9 +41,10 @@ class Banner(AppBase):
     )
 
     # History tracking
-    history = HistoricalRecords()
+    history = HistoricalRecords(table_name="core_historicalbanner")
 
     class Meta:
+        db_table = "core_banner"
         ordering = ["-modified"]
         verbose_name = "Banner"
         verbose_name_plural = "Banners"
@@ -78,3 +79,58 @@ class Banner(AppBase):
             raise models.ValidationError(
                 {"cta_text": "CTA text is required when CTA URL is provided."}
             )
+
+
+class ImpersonationLog(AppBase):
+    """Audit record of an admin impersonation session.
+
+    ``owner`` (inherited from :class:`~gyrinx.models.Owned`) is the impersonator —
+    the admin who started the session. ``target`` is the user who was impersonated.
+    ``created`` marks the start; ``ended_at`` / ``ended_reason`` are filled in when
+    the session stops (manually, on logout, on timeout, or when it is revoked
+    automatically because the admin lost privileges or the target went away).
+
+    This is an append-only audit log — like :class:`~n23.core.models.events.Event`
+    it does not declare ``HistoricalRecords``.
+    """
+
+    class EndedReason(models.TextChoices):
+        MANUAL = "manual", "Stopped manually"
+        LOGOUT = "logout", "Logged out"
+        EXPIRED = "expired", "Timed out"
+        REVOKED = "revoked", "Ended automatically"
+
+    target = models.ForeignKey(
+        "auth.User",
+        on_delete=models.SET_NULL,
+        null=True,
+        blank=True,
+        related_name="impersonated_by_sessions",
+        help_text="The user who was impersonated.",
+    )
+    ended_at = models.DateTimeField(
+        null=True,
+        blank=True,
+        help_text="When the impersonation session ended (null while active).",
+    )
+    ended_reason = models.CharField(
+        max_length=20,
+        choices=EndedReason.choices,
+        blank=True,
+        help_text="Why the impersonation session ended.",
+    )
+
+    class Meta:
+        db_table = "core_impersonationlog"
+        verbose_name = "impersonation log"
+        verbose_name_plural = "impersonation logs"
+        ordering = ["-created"]
+
+    @property
+    def impersonator(self):
+        """Alias for ``owner`` — the admin who started the session."""
+        return self.owner
+
+    def __str__(self):
+        state = "active" if self.ended_at is None else self.ended_reason
+        return f"{self.owner} → {self.target} ({state})"
