+++ n23/core/models/__init__.py
@@ -5,7 +5,6 @@ from .campaign import *  # noqa: F403
 from .crew import *  # noqa: F403
 from .events import *  # noqa: F403
 from .facts import *  # noqa: F403
-from .impersonation import *  # noqa: F403
 from .invitation import *  # noqa: F403
 from .list import *  # noqa: F403
 from .notification import *  # noqa: F403
@@ -16,5 +15,4 @@ from .pack import (  # noqa: F401
     CustomContentPackPermission,
 )
 from .print_config import *  # noqa: F403
-from .site import *  # noqa: F403
 from .upload import *  # noqa: F403
