+++ n23/core/models/crew.py
@@ -34,7 +34,7 @@ from django.core.exceptions import ValidationError
 from django.db import models
 from simple_history.models import HistoricalRecords
 
-from n23.core.models.base import AppBase
+from gyrinx.base_models import AppBase
 
 __all__ = [
     "Crew",
