+++ n23/core/models/list/assignment.py
@@ -22,7 +22,7 @@ from n23.content.models import (
     VirtualWeaponProfile,
 )
 from n23.core.models.facts import AssignmentFacts
-from n23.core.models.history_mixin import HistoryMixin
+from gyrinx.history_mixin import HistoryMixin
 from n23.core.models.list._common import preferred_equipment_list_override
 from gyrinx.models import Archived, Base
 from n23.models import format_cost_display
