+++ n23/core/models/list/campaign_state.py
@@ -13,7 +13,7 @@ from n23.content.models import (
     ContentSkillCategory,
     ContentStatlineTypeStat,
 )
-from n23.core.models.base import AppBase
+from gyrinx.base_models import AppBase
 from n23.core.models.list.fighter import ListFighter
 from n23.core.models.list.list import List
 from gyrinx.models import Archived, Base
