+++ n23/core/models/list/counter_spend.py
@@ -15,7 +15,7 @@ record itself.
 from django.db import models
 from simple_history.models import HistoricalRecords
 
-from n23.core.models.base import AppBase
+from gyrinx.base_models import AppBase
 from n23.core.models.list.fighter import ListFighter
