+++ n23/core/models/list/equipment_set.py
@@ -15,7 +15,7 @@ never persisted. See issue #1853.
 from django.db import models
 from simple_history.models import HistoricalRecords
 
-from n23.core.models.base import AppBase
+from gyrinx.base_models import AppBase
 
 
 class ListFighterEquipmentSet(AppBase):
