+++ n23/core/models/list/roll_result.py
@@ -11,7 +11,7 @@ mirroring how advancements work.
 from django.db import models
 from simple_history.models import HistoricalRecords
 
-from n23.core.models.base import AppBase
+from gyrinx.base_models import AppBase
 from n23.core.models.list.fighter import ListFighter
