+++ n23/core/models/notification.py
@@ -16,8 +16,8 @@ from django.db.models import Q
 from django.urls import reverse
 from django.utils import timezone
 
-from n23.core.models.base import AppBase
-from n23.core.models.history_aware_manager import HistoryAwareManager
+from gyrinx.base_models import AppBase
+from gyrinx.history_aware_manager import HistoryAwareManager
 
 logger = logging.getLogger(__name__)
