+++ n23/core/views/impersonation.py
@@ -20,7 +20,7 @@ from gyrinx.impersonation import (
     can_impersonate,
     can_impersonate_target,
 )
-from n23.core.models import ImpersonationLog
+from gyrinx.site.models import ImpersonationLog
 from n23.core.utils import safe_redirect
